The pre process of the parameters of the request functions

This function is of no particular use to the user, but might be useful for anyone who would like to dive deeper into the package development. This function is used to add new parameters to the final GET request.

add_para(new_arg, arg_name, args)

Arguments

new_arg
The new argument to be added.

arg_name
The name of the argument in the final url.

args
The current argument list

Value

A list of parameters used in get_response function.

Examples

# NOT RUN {
add_para(new_arg = players, arg_name = "players", args = args)
# }