Unique Heroes

A list of players and the unique heroes they have used

get_player_unique(patch = "7.19", split = "lan",
  after = "2011.01.01", before = "2018.07.12", duration = "0-100",
  leagues = NULL, tier = "1&2&3", valve = "does-not-matter",
  teams = NULL, players = NULL, roles = NULL)

Arguments

patch
Which patches are you interested in? From 6.74 to 7.20. (Connected by "&")

split
What type of event was it? Choose from "online", "lan" or "post-event".

after
The beginning date of the match

before
The last day of the period to search

duration
The time the game lasts (measured in minutes)

leagues
What league is it in? Check the serial number of the league on www.datdota.com.

tier
What tier of event is this? 1 = Premium; 2 = Professional; 3 = Semi-pro; (Connected by "&")

valve
Is this a Valve event? Choose from "yes", "no" or "does-not-matter".

teams
Teams to serch for. Check the serial number of the players on www.datdota.com. (Connected by "&". At least one must appear)

players
Players to search for. Check the serial number of the players on www.datdota.com. (Connected by "&". Can be null.)

roles
What role was it played? Choose from "core" or "support". (Connected by "&". Can be null.)

Value

A dota_api object containing the elements described in the details.

Details

A list will be returned that contains three elements. The content, the url and the response received from the api.

The content element of the list contains a sub list called data. Following informations are included: nickname, steamId, heroes (list), countUnique, gameCount.

steamID: The steam ID of the player.

nickname: The nickname of the player.

heroes: A list of unique heroes that the player have played.

countUnique: The number of unique heroes that the player have used.

gameCount: Total nuber of games.

Examples

# NOT RUN {
get_player_unique()
get_player_unique(patch = "7.19", split = "lan", after = "2011.01.01", before = "2018.07.12",
                  duration = "0-100", leagues = NULL, tier = "1&2&3", valve = "does-not-matter",
                  teams = "15", players = "106863163", roles = "core")
# }